decentralization admin

+++ mvc/model/categories.php
@@ -3,34 +3,46 @@
         if (!isset($_SESSION['users']['email'])) {
             header("location:login");
         }
+    
     class categories extends DB{
+        
         public function insert_cate(){
-            if (!empty($_POST)) {
-                $name   = $_POST['name'];
-                $slug   = $_POST['slug'];
-                if (!empty($name) && !empty($slug)){
-                    $sql = "INSERT INTO categories(id, cate_name, slug) VALUES (NULL, '$name', '$slug')";
-                    mysqli_query($this->con,$sql);
-                    header('Location: list_categories'); 
-                } else {
-                    $msg = "Please fill out the information!";
+            if($_SESSION['users']['email']=="admin@gmail.com"){
+                if (!empty($_POST)) {
+                    $name   = $_POST['name'];
+                    $slug   = $_POST['slug'];
+                    if (!empty($name) && !empty($slug)){
+                        $sql = "INSERT INTO categories(id, cate_name, slug) VALUES (NULL, '$name', '$slug')";
+                        mysqli_query($this->con,$sql);
+                        header('Location: list_categories'); 
+                    } else {
+                        $msg = "Please fill out the information!";
+                    }
                 }
-            }   
+            } else {
+                header("location:list_categories");
+            } 
 
         }
         public function get_category(){
 
             $sql ="SELECT * FROM categories";
             return mysqli_query($this->con,$sql);
+
         }
 
         public function delete_cate(){
-            if(!empty($_POST["delete"])){
-                $id = $_POST["delete"];
-                $sql =  "delete from categories where id= $id";
-                return mysqli_query($this->con,$sql);
-                echo "Success";
+            if($_SESSION['users']['email']=="admin@gmail.com"){
+                if(!empty($_POST["delete"])){
+                    $id = $_POST["delete"];
+                    $sql =  "delete from categories where id= $id";
+                    return mysqli_query($this->con,$sql);
+                    echo "Success";
+                }
+            }else{
+                header("location:list_categories");
             }
+            
         }
         public function get_cate_by_id(){
             if(!empty($_GET['url'])){
@@ -48,7 +60,8 @@
         }
         public function update_cate(){
             
-            if (!empty($_POST['update_cate']))
+            if($_SESSION['users']['email']=="admin@gmail.com"){
+                if (!empty($_POST['update_cate']))
                 {
                 unset($_GET["url"]);
                 $id_cate = $_POST['id']; 
@@ -59,6 +72,10 @@
                     header('Location: /baitap/home/list_categories'); 
                 }
                 unset($_GET['url']);
+            } else {
+                header("location:list_categories");
+            }
+            
         }
 
     }
